Android Clubbing App
An Android Application built using Android Studio. The Application
allows users to search for venues for information such as Bands/Djs
playing there. It holds the information in a remote mySQL database. The
application connects to the database and downloads it in JSON format and
displays it in the application for the user